Add lib to cli (deno) crate and add tools feature
We want to be able to call the cli code from within exograph so
need more than a binary only package.

The feature leaves out of a lot of the deno cli build related to the cli
executable, lsp etc, none of which we need to use the lib in exograph.

Running:

- cargo build
- cargo build --features tools

should both work at this point.