Fix issue with release build and zstd
The cli/build.rs has a dependency on zstd for release builds which
is an optional dependency behind the `tools` flag.

This adds the flag to the snapshot creation functions in the build
file to prevent a compiler error when the dependency is missing.

Also add some suppression of unused code warnings which is preferable
to adding `tools` to every unused import or variable.